begin refactoring to allow instances to be created and closed separately
